# Speed up Attribute instantiation.