Hello, it's me again.

This speeds up Attribute instantiation. 

Before:
```
> pyperf timeit --rigorous -g -s "from attr import Attribute, NOTHING" "Attribute(name='test', default=NOTHING, validator=None, repr=True, cmp=True, hash=True, init=True)"
........................................
5.58 us:  2 #########
5.63 us:  6 ############################
5.68 us:  5 #######################
5.72 us:  9 ##########################################
5.77 us: 11 ###################################################
5.81 us: 17 ###############################################################################
5.86 us:  7 #################################
5.91 us: 10 ##############################################
5.95 us: 15 ######################################################################
6.00 us: 13 ############################################################
6.05 us:  8 #####################################
6.09 us:  6 ############################
6.14 us:  2 #########
6.18 us:  2 #########
6.23 us:  2 #########
6.28 us:  1 #####
6.32 us:  2 #########
6.37 us:  0 |
6.41 us:  0 |
6.46 us:  0 |
6.51 us:  2 #########

Median +- std dev: 5.94 us +- 0.18 us
```

After:
```
> pyperf timeit --rigorous -g -s "from attr import Attribute, NOTHING" "Attribute(name='test', default=NOTHING, validator=None, repr=True, cmp=True, hash=True, init=True)"
........................................
2.20 us:  1 #####
2.22 us:  3 ################
2.23 us:  6 ################################
2.24 us:  9 ###############################################
2.26 us: 14 ##########################################################################
2.27 us: 14 ##########################################################################
2.28 us:  8 ##########################################
2.30 us:  8 ##########################################
2.31 us: 15 ###############################################################################
2.33 us:  9 ###############################################
2.34 us:  7 #####################################
2.35 us:  4 #####################
2.37 us:  3 ################
2.38 us:  5 ##########################
2.39 us:  6 ################################
2.41 us:  6 ################################
2.42 us:  1 #####
2.43 us:  0 |
2.45 us:  0 |
2.46 us:  0 |
2.48 us:  1 #####

Median +- std dev: 2.31 us +- 0.06 us
```

The gains are significant, and likely to help a lot of users. 🏎