@Tinche 2016-08-22
Ugh, Python 2 has slightly different error messages. I'll get around to it.