@Tinche 2016-08-22
Is it ok if I just remove the test_missing_argument test? `Attribute` is basically a normal class now, with minimal logic in the initializer. I could put in the Py2 error message but I feel like we're testing whether classes need to take all their `__init__` args at this point.